Extract reusable BaseToastMessages component
